Implement deriveVisibleDates function for enhanced date management in course creation and updates
- Introduced a new utility function, deriveVisibleDates, to calculate visible dates based on various parameters such as planning mode, visibility mode, and date exclusions/inclusions.
- Updated createCourse and updateCourse lambdas to utilize the deriveVisibleDates function for managing course dates effectively.
- Enhanced tests for createCourse to validate the correct calculation of visible dates based on the provided input.
- Refactored getCourses lambda to include the new date management logic, ensuring consistency across course-related functionalities.
